image-patch-extractor/requirements.txt
certifi==2021.10.8
cv==1.0.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
wincertstore==0.2
